pyproject.toml
[project]
name = "conversational-search"
version = "0.1.0"
description = "Conversational search assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db-dtypes>=1.5.0",
    "fastapi[standard]>=0.118.3",
    "gcsfs>=2026.1.0",
    "google-adk>=1.19.0",
    "google-auth>=2.43.0",
    "google-cloud-bigquery>=3.38.0",
    "httpx>=0.28.1",
    "litellm>=1.80.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyarrow>=23.0.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.3",
    "qdrant-client[fastembed]>=1.16.2",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "mypy>=1.19.0",
    "pandas-stubs>=2.3.3.251201",
    "pylint>=4.0.4",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.7",
    "ty>=0.0.1a29",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.mypy]
show_error_codes = true
ignore_missing_imports = false
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.isort]
profile = "black"

[tool.black]
line-length = 100
